bug fixed

+++ src/main/java/com/example/bio/controller/BioCategoryController.java
@@ -1,6 +1,7 @@
 package com.example.bio.controller;
 
 
+import cn.hutool.core.util.StrUtil;
 import com.baomidou.mybatisplus.core.conditions.update.UpdateWrapper;
 import com.example.bio.common.api.BaseController;
 import com.example.bio.common.api.Result;
@@ -54,6 +55,9 @@ public class BioCategoryController extends BaseController {
     @PreAuthorize("hasRole('ADMIN')")
     @DeleteMapping("/delete/{id}")
     public Result<?> deleteCategoryById(@PathVariable("id") String id) {
+        if (StrUtil.isBlank(id)) {
+            return fail("请输入正确的id");
+        }
         categoryService.removeById(id);
         return ok("删除成功");
     }
@@ -61,6 +65,9 @@ public class BioCategoryController extends BaseController {
     @PutMapping("/update/{id}")
     public Result<?> updateCategoryById(@PathVariable("id") String id,
                                         @RequestParam String categoryName) {
+        if (StrUtil.isBlank(id)) {
+            return fail("请输入正确的id");
+        }
         UpdateWrapper<BioCategory> updateWrapper = new UpdateWrapper<>();
         updateWrapper
                 .eq("id", id)
