+++ src/main/java/com/example/bio/controller/BioCommentController.java
@@ -1,12 +1,15 @@
 package com.example.bio.controller;
 
 
+import cn.hutool.core.util.StrUtil;
 import com.example.bio.common.api.BaseController;
 import com.example.bio.common.api.Result;
 import com.example.bio.common.domain.PageQueryParams;
 import com.example.bio.dto.CommentDto;
 import com.example.bio.model.BioComment;
 import com.example.bio.service.BioCommentService;
+import com.example.bio.util.ResponseUtil;
+import com.google.common.annotations.Beta;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -46,6 +49,9 @@ public class BioCommentController extends BaseController {
     @ApiOperation(value = "删除评论", notes = "级联删除，删除父评论则子评论全部删除,自传拥有者，评论拥有者可以删除")
     @DeleteMapping("deleteCommentById/{id}")
     public Result<?> deleteCommentById(@PathVariable("id") String id) {
+        if (StrUtil.isBlank(id)) {
+            return fail("请输入正确的id");
+        }
         commentService.deleteCommentById(id);
         return ok("删除成功");
     }
